Refactor XORTrialWithTrainingMetrics
