+++ harness/tests/experiment/pytorch/test_pytorch_trial.py
@@ -112,7 +112,7 @@ class TestPyTorchTrial:
         )
         controller.run()
 
-    def test_xor_training_metrics(self) -> None:
+    def test_training_metrics(self) -> None:
         def make_workloads() -> workload.Stream:
             trainer = utils.TrainAndValidate()
 
@@ -120,10 +120,10 @@ class TestPyTorchTrial:
             training_metrics, validation_metrics = trainer.result()
 
             for metrics in training_metrics:
-                assert "accuracy" in metrics
+                assert "mse" in metrics
 
         controller = utils.make_trial_controller_from_trial_implementation(
-            trial_class=pytorch_xor_model.XORTrialWithTrainingMetrics,
+            trial_class=pytorch_onevar_model.OneVarTrialWithTrainingMetrics,
             hparams=self.hparams,
             workloads=make_workloads(),
             trial_seed=self.trial_seed,
